+++ recipes/client.rb
@@ -49,7 +49,16 @@ edit_resource(:service, 'splunk') do
   start_command svc_command('start')
   restart_command svc_command('restart')
   status_command svc_command('status')
+  notifies :run, "execute[#{splunk_cmd} stop]", :before unless correct_runas_user?
   provider splunk_service_provider
+end
+
+# if the splunk daemon is running as root, executing a normal service restart or stop will fail if the boot
+# start script has been modified to execute splunk as a non-root user.
+# So, the splunk daemon must be run this way instead
+execute "#{splunk_cmd} stop" do
+  action :nothing
+  not_if { node['splunk']['server']['runasroot'] == true }
 end
 
 directory "#{splunk_dir}/etc/system/local" do